[CRM-2][CRM-7] Create build process - developer machine setup
Mainly the stuff to download nuget package for PowerApps CLI and use Chocolatey to install it - but also the .Net install stuff that we may need for later

Ignore the stuff in the setup folder for now. It will be required later (and will need updating/modifying) but at the moment, is just a placeholder.