[CRM-2] Support for initialising a Dynamics 365 CRM development environment prior to applying changes
Contains:
[CRM-6] Create build process - package solution
[CRM-8] Create build process - deploy solution and manage target environment solution(s)

At the moment, I have set this up to deal with One (1) solution only, because

A> That's all we need to deal with at the moment
B> Not sure if the strategy going forward will be to have multiple solutions in a single repo. It certainly complicates the deploy scripts. Initially we said yes, now not so sure.